testing lazygit

+++ lib/libft/include/libft.h
@@ -52,7 +52,7 @@ int					ft_memcmp(const void *s1, const void *s2, size_t n);
 char				*ft_strnstr(const char *big, const char *little,
 						size_t len);
 int					ft_atoi(const char *nptr);
-double				ft_atof(const char *nptr);
+float				ft_atof(const char *nptr);
 bool				ft_atoib(const char *nptr);
 long				ft_atol(const char *nptr);
 void				*ft_calloc(size_t nmemb, size_t size);
